LichTrinhNLP/requirement.txt
underthesea
python-dateutil
scikit-learn
tk
